drop glpi_networkportmigrations only if presents

+++ install/migrations/update_10.0.x_to_11.0.0/networkportmigrations.php
@@ -36,4 +36,6 @@
 /**
  * @var \Migration $migration
  */
-$migration->dropTable('glpi_networkportmigrations');
+if ($DB->tableExists('glpi_networkportmigrations')) {
+    $migration->dropTable('glpi_networkportmigrations');
+}
